requirements.txt
# local package
-e .

# external requirements

numpy==1.19.5
pandas==1.4.1
scikit-learn==1.0.2
matplotlib==3.5.1
seaborn==0.11.2
cloudpickle==2.0.0
optuna==2.10.0
missingno==0.5.0
boruta==0.3
mlflow
hydra-core
flask
mlxtend==0.19.0
yellowbrick==1.4